TEST: Fix the stress tests using the XrdStress class to properly use the "-c" flag       which means that jobs (threads/procs) do the operations (rd/wr) on the same       files concurrently FUSE: Save nicely formatted file paths in the cache ( remove duplicate "//" ) CMAKE: Drop FindXrdCl.cmake as XrdCl is now part of the standard XRootD distribution

+++ fst/layout/XrdFileIo.cc
@@ -218,7 +218,7 @@ XrdFileIo::Read( XrdSfsFileOffset offset,
 	SimpleHandler* sh = iter->second->handler;
 
 	if ( sh->WaitOK() ) {
-	  read_length = ( length < mBlocksize ) ? length : mBlocksize;
+	  read_length = ( length < (int64_t) mBlocksize ) ? length : mBlocksize;
 	  pBuff = static_cast<char*>( memcpy( pBuff,
 					      iter->second->buffer,
 					      read_length ) );
